refactor: 영화 API 호출을 위한 fetchMovies 함수 분리
